Revert "Deprecate Slick and move Projected"
